requirements.txt
aniso8601==9.0.1
asttokens==2.4.0
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
comm==0.1.4
cryptography==41.0.4
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
idna==3.4
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.4.0
jupyter_core==5.4.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mysql-connector-python==8.1.0
nest-asyncio==1.5.8
numpy==1.26.1
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
panda==0.3.1
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.6
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.21
stack-data==0.6.3
tornado==6.3.3
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.8
Werkzeug==2.3.7
